chore(deps): upgrade tree-sitter, notify, rusqlite, rmcp to latest
- tree-sitter: 0.24 → 0.26; grammar crates 0.23 → 0.25 (python, js, go) / 0.24 (rust)
- notify: 7 → 8; notify-debouncer-mini: 0.5 → 0.7
- rusqlite: 0.31 → 0.39
- rmcp: 0.5 → 1.2
- rust-version: 1.70 → 1.77 (required by notify 8)

Fix breaking changes:
- node.child(i)/named_child(n) now require u32 — cast usize loop vars with `as u32`
  in go.rs, js_shared.rs, python.rs, rust_lang.rs
- rmcp: Parameters moved to handler::server::wrapper::Parameters
- rmcp: ServerInfo/Implementation are #[non_exhaustive] — migrate to builder pattern

+++ src/languages/go.rs
@@ -197,7 +197,7 @@ fn extract_receiver_type(node: Node, source: &str) -> Option<String> {
             // Fallback: last named child
             let count = child.named_child_count();
             if count > 0 {
-                if let Some(last) = child.named_child(count - 1) {
+                if let Some(last) = child.named_child((count - 1) as u32) {
                     return Some(extract_type_name(last, source));
                 }
             }
